service-py/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
redis==5.0.1
kafka-python==2.0.2
